refactor: restructure tracks feature components into logical groups
